# Refactor Hilbert values computation