--- a/features/testbot/bearing_param.feature
+++ b/features/testbot/bearing_param.feature
@@ -48,7 +48,7 @@ Feature: Bearing parameter
     Scenario: Testbot - Initial bearing on split way
         Given the node map
            """
-           g d         1         c f
+           g d 2       1         c f
            h a         0         b e
            """
 
@@ -64,21 +64,17 @@ Feature: Bearing parameter
             | ha    | yes    |
 
         When I route I should get
-            | from | to | bearings | route          | bearing                             |
-            | 0    | b  | 10 10    | bc,bc          | 0->0,0->0                           |
-            | 0    | b  | 90 90    | ab,ab          | 0->90,90->0                         |
-            # The returned bearing is wrong here, it's based on the snapped
-            # coordinates, not the acutal edge bearing.  This should be
-            # fixed one day, but it's only a problem when we snap two vias
-            # to the same point - DP
-            #| 0    | b  | 170 170  | da          | 180                                   |
-            #| 0    | b  | 189 189  | da          | 180                                   |
-            | 0    | 1  | 90 270   | ab,bc,cd,cd    | 0->90,90->0,0->270,270->0           |
-            | 1    | d  | 10 10    | bc,bc          | 0->0,0->0                           |
-            | 1    | d  | 90 90    | ab,bc,cd,da,da | 0->90,90->0,0->270,270->180,180->0  |
-            | 1    | 0  | 189 189  | da,da          | 0->180,180->0                       |
-            | 1    | d  | 270 270  | cd,cd          | 0->270,270->0                       |
-            | 1    | d  | 349 349  |                |                                     |
+            | from | to | bearings | route             | bearing                                   |
+            | 0    | b  | 10 10    | bc,bc             | 0->0,0->0                                 |
+            | 0    | b  | 90 90    | ab,ab             | 0->90,90->0                               |
+            | 0    | b  | 170 170  | da,da             | 0->0,0->0                                 |
+            | 0    | b  | 189 189  | da,da             | 0->0,0->0                                 |
+            | 0    | 1  | 90 270   | ab,bc,cd,cd       | 0->90,90->0,0->270,270->0                 |
+            | 1    | 2  | 10 10    | bc,bc             | 0->0,0->0                                 |
+            | 1    | 2  | 90 90    | ab,bc,cd,da,ab,ab | 0->90,90->0,0->270,270->180,180->90,90->0 |
+            | 1    | 0  | 189 189  | da,da             | 0->180,180->0                             |
+            | 1    | 2  | 270 270  | cd,cd             | 0->270,270->0                             |
+            | 1    | 2  | 349 349  |                   |                                           |
 
     Scenario: Testbot - Initial bearing in all direction
         Given the node map