# Issue

The current Hilbert codes is an SFC but not the original Hilbert curve #3352. The PR fixes Hilbert code computation. There are 6 failing test cases, that failing due to reordering of edges in the R-tree. For example, features/testbot/compression.feature:7 returns phantom node that uses 3 segments instead of 4, so the small adjustment of 0.11 meters is not added. Other tie-breaking cases look not so trivial and must be clarified first.

## Tasklist
 - [x] refactor implementation of Hilbert linear codes
 - [x] add regression / cucumber cases (see docs/testing.md)
 - [x] clarify failing features/car/maxspeed.feature:81
 - [x] clarify failing features/car/maxspeed.feature:100
 - [x] clarify failing features/car/summaries.feature:9
 - [x] clarify failing features/car/traffic_turn_penalties.feature:105
 - [x] clarify failing features/testbot/bearing_param.feature:48
 - [x] review
 - [x] adjust for comments

## Requirements / Relations
[Arndt, Jörg. Matters Computational Ideas, Algorithms, Source Code, 2010. p86](http://www.jjj.de/fxt/fxtbook.pdf)
http://home.pipeline.com/~hbaker1/hakmem/topology.html#item115