@danpat 2016-11-22
Quick comparison using `rtree-bench` with a small dataset (`monaco-latest.osm`):

Before this change:
```
Running raw RTree queries (1 result) with 10000 coordinates: Took 14.0459 seconds (14045.9ms)  ->  1.40459 ms/query (14045.9ms)
Running raw RTree queries (10 results) with 10000 coordinates: Took 14.3965 seconds (14396.5ms)  ->  1.43965 ms/query (14396.5ms)
```

After:
```
Running raw RTree queries (1 result) with 10000 coordinates: Took 13.4529 seconds (13452.9ms)  ->  1.34529 ms/query (13452.9ms)
Running raw RTree queries (10 results) with 10000 coordinates: Took 13.6612 seconds (13661.2ms)  ->  1.36612 ms/query (13661.2ms)
```

Looks like the corrected curve shape could give us a modest performance boost due to improved cache locality.