@oxidase 2016-11-27
Tests features/testbot/compression.feature:7,  failing features/car/maxspeed.feature:81,  failing features/car/maxspeed.feature:100 use exact values for expectations, but  other ordering could return different edges for phantom nodes and due to some round-off errors introduce 0.11 meters segments that correspond to 0.000001 coordinate precision. Expectation in tests are update to the new values that are still ordering dependent.