@oxidase 2016-11-27
Tests features/car/summaries.feature:9, features/car/traffic_turn_penalties.feature:105 fail due to 0-length segments in snapped phantom nodes. 
Must be fixed in #2287