@oxidase 2016-11-27
Test features/testbot/bearing_param.feature:48 fail due to different components of the source and the target phantom nodes.

It should not be in real cases, so i would suggest to remove `ha` edge, but just commented out the test row